requirements.txt
fastapi>=0.104.1,<0.115.0
uvicorn[standard]>=0.24.0,<0.32.0
python-multipart>=0.0.6,<0.1.0
pydantic>=2.5.0,<3.0.0
PyMuPDF>=1.23.9,<1.25.0
filetype>=1.2.0,<2.0.0
aiofiles>=23.2.1,<25.0.0
websockets>=13.0,<15.0
google-genai==1.32.0
python-dotenv>=1.0.0,<2.0.0
pydantic-settings>=2.0.0,<3.0.0